% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffConcTime.R
\name{caffConcTime}
\alias{caffConcTime}
\title{Create a dataset of the concentration-time curve of single oral administration of caffeine}
\usage{
caffConcTime(Weight, Dose, N = 20)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of concentration and time of simulated subjects
}
\description{
\code{caffConcTime} will create a dataset of the concentration-time curve
}
\examples{
caffConcTime(Weight = 20, Dose = 200, N = 20)
caffConcTime(20, 200)
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
