% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffConcTime.R
\name{caffConcTimeMulti}
\alias{caffConcTimeMulti}
\title{Create a dataset of the concentration-time curve of multiple dosing of caffeine}
\usage{
caffConcTimeMulti(Weight, Dose, N = 20, Tau = 8, Repeat = 4)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}

\item{Tau}{The interval of multiple dosing (hour)}

\item{Repeat}{The number of dosing}
}
\value{
The dataset of concentration and time of simulated subjects of multiple dosing
}
\description{
\code{caffConcTimeMulti} will create a dataset of the concentration-time curve of multiple oral administrations of caffeine
}
\examples{
caffConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4)
caffConcTimeMulti(20, 200)
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
