% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{caffDataset}
\alias{caffDataset}
\title{Create a dataset for simulation of single dose of caffeine}
\usage{
caffDataset(Weight, Dose, N = 20)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal
}
\description{
\code{caffDataset} will create a dataset for simulation of single dose of caffeine
}
\examples{
caffDataset(Weight = 20, Dose = 200, N = 20)
caffDataset(20,500)
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
