% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffPkparam.R
\name{caffDescstat}
\alias{caffDescstat}
\title{Calculate descriptive statistics of simulated PK parameters}
\usage{
caffDescstat(caffPkparam)
}
\value{
The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal
}
\description{
\code{caffDescstat} will calculate descriptive statistics of simulated PK parameters
}
\examples{
caffDescstat(caffPkparam(20,500))
caffDescstat(caffPkparamMulti(20,500))
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
