% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffPkparam.R
\name{caffPkparam}
\alias{caffPkparam}
\title{Create a dataset of pharmacokinetic parameters of single oral dosing of caffeine}
\usage{
caffPkparam(Weight, Dose, N = 20)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal
}
\description{
\code{caffPkparam} will create a dataset for simulation of single dose of caffeine
}
\examples{
caffPkparam(Weight = 20, Dose = 200, N = 20)
caffPkparam(20,500)
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
