% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffPkparam.R
\name{caffPkparamMulti}
\alias{caffPkparamMulti}
\title{Create a dataset for simulation of multiple dose of caffeine}
\usage{
caffPkparamMulti(Weight, Dose, N = 20, Tau = 8)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of multiple caffeine (mg)}

\item{N}{The number of simulated subjects}

\item{Tau}{The interval of multiple dosing (hour)}
}
\value{
The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal
}
\description{
\code{caffPkparamMulti} will create a dataset for simulation of multiple dose of caffeine
}
\examples{
caffPkparamMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)
caffPkparamMulti(20,500)
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
