% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{caffPlot}
\alias{caffPlot}
\title{Create concentration-time curve after single dose of caffeine}
\usage{
caffPlot(caffConcTime, log = FALSE)
}
\arguments{
\item{log}{y axis log}

\item{ConcTime}{Concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}
}
\value{
The concentration-time curve
}
\description{
\code{Plot} will create concentration-time curve after single dose of caffeine
}
\examples{
caffPlot(caffConcTime(Weight = 20, Dose = 200, N = 20))
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
