% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{caffPlotMulti}
\alias{caffPlotMulti}
\title{Create concentration-time curve after multiple doses of caffeine}
\usage{
caffPlotMulti(caffConcTimeMulti, log = FALSE)
}
\arguments{
\item{log}{y axis log}

\item{ConcTimeMulti}{Concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}
}
\value{
The concentration-time curve
}
\description{
\code{PlotMulti} will create concentration-time curve after multiple doses of caffeine
}
\examples{
caffPlotMulti(ConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4))
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
