% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffShiny.R
\name{caffShiny}
\alias{caffShiny}
\title{Run shiny app to interactively simulate plasma caffeine concentration.}
\usage{
caffShiny()
}
\value{
shiny app
}
\description{
\code{caffShiny} will run an internal shiny app \code{Caffeine Concentration Predictor} in order to interactively simulate plasma caffeine concentration.
}
\seealso{
\url{https://asan.shinyapps.io/caff/}
}
