% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffsim.R
\name{caffsim}
\alias{caffsim}
\title{Compare variable names between CRF and data dictionary}
\usage{
caffsim(CRFcsv, Dictionaryxlsx)
}
\arguments{
\item{CRFcsv}{A filename of CRF csv file, exported from PDFCRF}

\item{Dictionaryxlsx}{A filename of data dictionary xlsx file}
}
\value{
List of output data of comparison of variables between a CRF-derived csv file and a data dictionary
}
\description{
\code{caffsim} uses CRF-derived csv file and a data dictionary excel to compare
variable names of each file
}
\examples{
\dontrun{
caffsim(CRFcsv = "foo.csv", Dictionaryxlsx = "foo.xlsx")
}
}
