% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcTime.R
\name{ConcTime}
\alias{ConcTime}
\title{Create a dataset of the concentration-time curve}
\usage{
ConcTime(Weight, Dose, N = 20)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of concentration and time of simulated subjects
}
\description{
\code{ConcTime} will create a dataset of the concentration-time curve
}
\examples{
ConcTime(Weight = 20, Dose = 200, N = 20)
ConcTime(20, 200)
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
