% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcTime.R
\name{ConcTimeMulti}
\alias{ConcTimeMulti}
\title{Create a dataset of the concentration-time curve of multiple dosing}
\usage{
ConcTimeMulti(Weight, Dose, N = 20, Tau = 8, Repeat = 4)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}

\item{Tau}{The interval of multiple dosing (hour)}

\item{Repeat}{The number of dosing}
}
\value{
The dataset of concentration and time of simulated subjects of multiple dosing
}
\description{
\code{ConcTimeMulti} will create a dataset of the concentration-time curve of multiple dosing
}
\examples{
ConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4)
ConcTimeMulti(20, 200)
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
