% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{Create a dataset for simulation of single dose of caffeine}
\usage{
Dataset(Weight, Dose, N)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single caffeine (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal
}
\description{
\code{Dataset} will create a dataset for simulation of single dose of caffeine
}
\examples{
Dataset(Weight = 20, Dose = 200, N = 20)
Dataset(20,500)
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
