% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetMulti.R
\name{DatasetMulti}
\alias{DatasetMulti}
\title{Create a dataset for simulation of multiple dose of caffeine}
\usage{
DatasetMulti(Weight, Dose, N = 20, Tau = 24)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of multiple caffeine (mg)}

\item{N}{The number of simulated subjects}

\item{Tau}{The interval of multiple dosing (hour)}
}
\value{
The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal
}
\description{
\code{DatasetMulti} will create a dataset for simulation of multiple dose of caffeine
}
\examples{
DatasetMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)
DatasetMulti(20,500)
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
