% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\title{Create concentration-time curve after single dose of caffeine}
\usage{
Plot(ConcTime, log = FALSE)
}
\arguments{
\item{ConcTime}{Concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}

\item{log}{y axis log}
}
\value{
The concentration-time curve
}
\description{
\code{Plot} will create concentration-time curve after single dose of caffeine
}
\examples{
Plot(ConcTime(Weight = 20, Dose = 200, N = 20))
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
