% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMulti.R
\name{PlotMulti}
\alias{PlotMulti}
\title{Create concentration-time curve after multiple doses of caffeine}
\usage{
PlotMulti(ggsuper)
}
\arguments{
\item{ConcTime}{Concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}
}
\value{
The concentration-time curve
}
\description{
\code{PlotMulti} will create concentration-time curve after multiple doses of caffeine
}
\examples{
PlotMulti(ConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4))
}
\seealso{
\url{http://asancpt.github.io/CaffeineEdison}
}
