% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffDescstat.R
\name{caffDescstat}
\alias{caffDescstat}
\title{Calculate descriptive statistics of simulated PK parameters}
\usage{
caffDescstat(caffPkparamData)
}
\arguments{
\item{caffPkparamData}{data frame generated by caffPkparam function}
}
\value{
The descriptive statistics of pharmacokinetic parameters
}
\description{
\code{caffDescstat} will calculate descriptive statistics of simulated PK parameters
}
\examples{
caffDescstat(caffPkparam(20,500))
caffDescstat(caffPkparamMulti(20,500))
caffDescExample <- cbind(caffDescstat(caffPkparam(20,500)), 
                         caffDescstat(caffPkparam(50,500))[,2])
colnames(caffDescExample)[2:3] <- c('20 kg', '50 kg')      
caffDescExample
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
