% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{caffPlotMulti}
\alias{caffPlotMulti}
\title{Create concentration-time curve after multiple doses of caffeine}
\usage{
caffPlotMulti(caffConcTimeMultiData, log = FALSE)
}
\arguments{
\item{caffConcTimeMultiData}{Concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}

\item{log}{y axis log}
}
\value{
The concentration-time curve
}
\description{
\code{caffPlotMulti} will create concentration-time curve after multiple doses of caffeine
}
\examples{
\donttest{caffPlotMulti(ConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4))}
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
