% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caffShiny.R
\name{caffShiny}
\alias{caffShiny}
\title{Run Shiny app to interactively simulate single and multiple dosing for plasma caffeine concentration}
\usage{
caffShiny()
}
\description{
\code{caffShiny} runs an internal shiny app \code{Caffeine Concentration Predictor} in order to interactively simulate plasma caffeine concentration.
}
\seealso{
\url{https://asan.shinyapps.io/caff/}
}
