% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitTable}
\alias{UnitTable}
\title{Unit data of PK parameters}
\format{A data frame with 16 rows and 2 variables:
\describe{
  \item{Parameters}{Abbreviated pharmacokinetic parameters}
  \item{Parameter}{Pharmacokinetic parameters in full name}
}}
\usage{
UnitTable
}
\description{
A dataset containing information regarding unit data of pharmacokinetic parameters
}
\seealso{
\url{https://asancpt.github.io/caffsim}
}
\keyword{datasets}
